feat: 添加预置字典分组功能
- 在 DictGroup 实体中新增 `preset` 字段，表示字典分组是否为预置。
- 更新 DictService 的删除逻辑，防止删除预置字典分组。
- 修改相关请求和响应数据模型，包含 `preset` 字段的定义。
- 更新数据库表结构，添加 `preset` 列并设置默认值。
- 更新 OpenAPI 文档，反映新增的 `preset` 字段。

+++ mahu-admin/mahu-admin-infra/src/main/java/cool/houge/mahu/admin/sys/service/DictService.java
@@ -57,6 +57,10 @@ public class DictService {
     @Transactional
     public void deleteById(String id) {
         var bean = obtainById(id);
+        if (bean.isPreset()) {
+            throw new BizCodeException(BizCodes.INVALID_ARGUMENT, "预置字典分组不可删除");
+        }
+
         dictGroupRepository.delete(bean);
         log.info("删除数据字典 {}", bean);
     }
